% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_correction_vectorized.R
\name{baseline_correction_vectorized}
\alias{baseline_correction_vectorized}
\title{Applies rubberband baseline correction over entire vector}
\usage{
baseline_correction_vectorized(x)
}
\arguments{
\item{x}{a vector containing measurment data}
}
\value{
a baseline corrected data vector
}
\description{
Applies rubberband baseline correction over entire vector
}
\examples{
baseline_correction_vectorized(UV_data)
}
