% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_SEC.R
\name{Build_SEC}
\alias{Build_SEC}
\title{Composes SEC matrices according to UV Vis and CV data}
\usage{
Build_SEC(y, x, scan = 2)
}
\arguments{
\item{y}{a list of CV matrices obtained from Load_CV()}

\item{x}{a matrix of UV experiments, obtained from Load_UV()}

\item{scan}{which CV scan to take from CV}
}
\value{
a list of SEC matrices for the oxidative and reductive scan respectively
}
\description{
Composes SEC matrices according to UV Vis and CV data
}
\examples{
Build_SEC(CV, UC, scan = 2)
}
