% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_seconds_ocean_view.R
\name{convert_to_seconds_ocean_view}
\alias{convert_to_seconds_ocean_view}
\title{Converts Ocean View file-timestamps to seconds}
\usage{
convert_to_seconds_ocean_view(file_list)
}
\arguments{
\item{file_list}{a list containing the file names ending in a time stamp with the format hh-mm-ss-mss}
}
\value{
the time difference in seconds between the each element and the earliest time stamp
}
\description{
Converts Ocean View file-timestamps to seconds
}
\examples{
convert_to_seconds_ocean_view(file_list)
}
