% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sparse_matrix.R
\name{create_sparse_matrix}
\alias{create_sparse_matrix}
\title{Creat sparse matrix with extrema information}
\usage{
create_sparse_matrix(x, nested_list)
}
\arguments{
\item{x}{a matrix, or vector, containing measurment data in each row}

\item{nested_list}{a list of extreme ponit indicies obtained from rowwise applying find_extreme_points}
}
\value{
a sparse filter matrix with extrema
}
\description{
Creat sparse matrix with extrema information
}
\examples{
create_sparse_matrix(x, peak_list)
}
