% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_df_to_matrix.R
\name{cv_df_to_matrix}
\alias{cv_df_to_matrix}
\title{Convert a data frame with CV data to a matrix}
\usage{
cv_df_to_matrix(x)
}
\arguments{
\item{x}{a data frame with cyclic voltammetry data}
}
\value{
a matrix containing the values for each experiment in its rows, per index, per scan.
}
\description{
Convert a data frame with CV data to a matrix
}
\examples{
cv_df_to_matrix(x, index == 2, scan == "reductive")
}
