% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclic_voltammetry_analysis.R
\name{cyclic_voltammetry_analysis}
\alias{cyclic_voltammetry_analysis}
\title{Performs peak analysis and baseline correction on cyclic voltammetry data}
\usage{
cyclic_voltammetry_analysis(x, index = 2)
}
\arguments{
\item{x}{List of matrices containing cyclic voltammetry data, obtained from Load_CV()}

\item{index}{Which scan number should be analysed. Can also be iterated over.}
}
\value{
A nested list containing peak information
}
\description{
Performs peak analysis and baseline correction on cyclic voltammetry data
}
\examples{
cyclic_voltammetry_analysis(x)
}
