% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample}
\alias{downsample}
\title{Downsamples a vector to a specified size.}
\usage{
downsample(vector, n)
}
\arguments{
\item{vector}{A vector of values to downsample.}

\item{n}{desired number of points.}
}
\value{
a vector of n equally spaced points
}
\description{
Downsamples a vector to a specified size.
}
\examples{
downsample(UV[, 1], n = 500)
}
