% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_extreme_points.R
\name{find_extreme_points}
\alias{find_extreme_points}
\title{Looks for extreme points in a time series via zero intersections in the first and second derivative.}
\usage{
find_extreme_points(x, span = 0.05, epsilon_scale = 500)
}
\arguments{
\item{x}{a time series vector}

\item{span}{argument passed to LOESS smoothing function, necessary to avoid "error multiplication" on noisy data.}

\item{epsilon_scale}{scale of epsilon region with respect to smallest absolute value in the time series, used to find estimate zero intercepts.}
}
\value{
a matrix containing extreme location information.
}
\description{
Looks for extreme points in a time series via zero intersections in the first and second derivative.
}
\examples{
dir <- "/path/to/CV"
data <- Load_CV(dir)
peaks <- find_extreme_points(data$V2, type = "peak")
valleys <- find_extreme_points(data$V2, type = "valley")
saddles <- find_extreme_points(data$V2, type = "saddle")
inflections <- find_extreme_points(data$V2, type = "inflection")
}
