% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_zero_crossings.R
\name{find_zero_crossings}
\alias{find_zero_crossings}
\title{Find closest value in zero crossing}
\usage{
find_zero_crossings(vector)
}
\arguments{
\item{vector}{a vector wit zero crossings}
}
\value{
a vector of indices where zero crossing happens
}
\description{
Find closest value in zero crossing
}
\examples{
find_zero_crossings(x)
}
