% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_baseline_peak_wise.R
\name{fit_baseline}
\alias{fit_baseline}
\title{Fit a baseline with convex hull and calculate true peak height for each peak}
\usage{
fit_baseline(x, span)
}
\arguments{
\item{x}{a sparse matrix or vector with peak information}

\item{span}{argument passed to LOESS smoother, the smaller the less smoothing. Warning values above 0.25 introduce major distortions to the data.}
}
\value{
a nested list with peak information
}
\description{
Fit a baseline with convex hull and calculate true peak height for each peak
}
\examples{
fit_baseline(x)
}
