% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_baseline_peak_wise.R
\name{fit_baseline_peak_wise}
\alias{fit_baseline_peak_wise}
\title{Fit a baseline with convex hull and calculate true peak height for each peak}
\usage{
fit_baseline_peak_wise(x, span, invert = FALSE)
}
\arguments{
\item{x}{A sparse matrix or vector with peak information}

\item{span}{Argument passed to LOESS smoother, the smaller the less smoothing. Warning values above 0.25 introduce major distortions to the data.}

\item{invert}{Logical, should data be inverted before and after peak analysis? Necessary for reductive CV scans.}
}
\value{
a nested list with peak information
}
\description{
Fit a baseline with convex hull and calculate true peak height for each peak
}
\examples{
fit_baseline(x)
}
