% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cv_data.R
\name{format_cv_data}
\alias{format_cv_data}
\title{Converts data frames with CV data to nested list of matrices.}
\usage{
format_cv_data(x)
}
\arguments{
\item{x}{A data frame containing CV data obtained from Load_CV()}
}
\value{
a nested list of matrices with measurement information strode row wise.
}
\description{
Converts data frames with CV data to nested list of matrices.
}
\examples{
format_cv_data(x)
}
