% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peak_info.R
\name{get_peak_info}
\alias{get_peak_info}
\title{get peak information from sparse matrix with peak values}
\usage{
get_peak_info(x)
}
\arguments{
\item{x}{a sparse matrix congaing peak values}
}
\value{
a nested list of peak information for each row of x
}
\description{
get peak information from sparse matrix with peak values
}
\examples{
get_peak_info(x)
}
