% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_monotonicity.R
\name{label_monotonicity}
\alias{label_monotonicity}
\title{Label monotonicity of cyclic voltammetry experiments to allow for separation of oxidative and reductive scan.}
\usage{
label_monotonicity(vec)
}
\arguments{
\item{vec}{a vector containing potential values of a cyclic voltammetry experiment}
}
\value{
a vector contining the scan direction
}
\description{
Label monotonicity of cyclic voltammetry experiments to allow for separation of oxidative and reductive scan.
}
\examples{
label_monotonicity(x)
}
