% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_CV.R
\name{Load_CV}
\alias{Load_CV}
\title{Load Cyclic Voltammetry data}
\usage{
Load_CV(dir, starting_row_number = 0, decimal_mark = ".", sep = "\\t")
}
\arguments{
\item{dir}{directory containing cyclic voltammetry data as .txt files recorded on CHI or Metrohm Instruments.}

\item{starting_row_number}{Skips rows if the file has a header. Choose the line where the numeric data starts, if text is included the function breaks. If no header nor column names are in the file leave this option blank. Default is 0}

\item{decimal_mark}{Decimal separator used in the data, is changed to ".".}

\item{sep}{Column separator used by the file.}
}
\value{
a matrix, with spectra as rows, row indices as relative acquisition time and column indices as wavelength.
}
\description{
Load Cyclic Voltammetry data
}
\examples{
dir <- "/path/to/cv/directory/"
CV_data <- Load_CV(dir)
}
