% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_Spectra.R
\name{Load_Spectra}
\alias{Load_Spectra}
\title{Load multiple txt files with UV-Vis raw data with wavelength and absorption values as columns and timestamps in the file names.}
\usage{
Load_Spectra(dir, starting_row_number = 0, decimal_mark = ".", sep = "\\t")
}
\arguments{
\item{dir}{Directory where the files are stored. Important!: Path must be separated by / not by \ as is default in the german version of windows!}

\item{starting_row_number}{Skips rows if the file has a header. Choose the line where the numeric data starts, if text is included the function breaks. If no header nor column names are in the file leave this option blank. Default is 0}

\item{decimal_mark}{Decimal separator used in the data, is changed to ".".}

\item{sep}{Column separator used by the file.}
}
\value{
matrix, with spectra as rows, row indices as relative acquisition time and column indices as wavelength.
}
\description{
Load multiple txt files with UV-Vis raw data with wavelength and absorption values as columns and timestamps in the file names.
}
\examples{
dir <- "/path/to/UV/data"
data <- Load_Spectra(dir, starting_row_number = 15, decimal_mark = ".", sep = "\t")
}
