% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_CV.R
\name{Plot_CV}
\alias{Plot_CV}
\title{Plot cyclic voltammetry data}
\usage{
Plot_CV(
  x,
  peak_info_cv,
  scan = 2,
  font = "Montserrat",
  col = "red",
  alpha = 0.7,
  show_peaks = FALSE
)
}
\arguments{
\item{x}{a list of matrices containing CV data, obtained from Load_CV().}

\item{peak_info_cv}{a list of peaks obtained by cyclic_voltammetry_analysis()}

\item{scan}{number of scan to be plotted.}

\item{font}{Font used in the plot.}

\item{col}{Color of CV curve.}

\item{alpha}{Alpha value of the curve.}

\item{show_peaks}{logical, should peaks be displayed?}
}
\value{
A Plot of a cyclic voltammetry experiment
}
\description{
Plot cyclic voltammetry data
}
\examples{
Plot_CV(x, scan = 2, font = "Montserrat", col ="red", alpha = 0.7, show_peaks = FALSE)
}
