% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_UV_range.R
\name{Plot_UV_range}
\alias{Plot_UV_range}
\title{Plot a range of UV spectra from a spectra matrix object.}
\usage{
Plot_UV_range(x, range = NULL, ROI = NULL)
}
\arguments{
\item{x}{a matrix with row wise spectra.}

\item{range}{a range of spectra to be selected}
}
\value{
a range of plotted UV-Vis spectra.
}
\description{
Plot a range of UV spectra from a spectra matrix object.
}
\examples{
Plot_UV_range(x, 20:50)
}
