% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorized_derivative.R
\name{vectorized_derivative}
\alias{vectorized_derivative}
\title{Take derivative with LOESS smooth before and after}
\usage{
vectorized_derivative(x, span = 0.05)
}
\arguments{
\item{x}{a vector, time series.}

\item{span}{argument passed to LOESS smoother, the smaller, the less smoothing and the slower.}
}
\value{
a vector of the smoothed numerical derivative of a time series.
}
\description{
Take derivative with LOESS smooth before and after
}
\examples{
vectorized_derivative(x, span = 0.045)
}
