% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorized_LOESS.R
\name{vectorized_LOESS}
\alias{vectorized_LOESS}
\title{Vectorized LOESS smoothing.}
\usage{
vectorized_LOESS(x, span = 0.05)
}
\arguments{
\item{x}{a vector to be smoothed.}

\item{span}{argument passed to LOESS.}
}
\value{
a smoothed vector.
}
\description{
Vectorized LOESS smoothing.
}
\examples{
vectorized_LOESS(x, span = 0.034)
}
