% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sparse_matrix.R
\name{create_sparse_matrix}
\alias{create_sparse_matrix}
\title{Create sparse matrix with extreme information}
\usage{
create_sparse_matrix(x, nested_list)
}
\arguments{
\item{x}{a matrix, or vector, containing measurement data in each row}

\item{nested_list}{a list of extreme point indices obtained from row wise applying find_extreme_points}
}
\value{
a sparse filter matrix with extreme
}
\description{
Create sparse matrix with extreme information
}
\examples{
create_sparse_matrix(x, peak_list)
}
