% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_points_sparse.R
\name{extreme_points_sparse}
\alias{extreme_points_sparse}
\title{Finds kind and location of extrema and converts the to a sparse matrix, with measurment number as row indicies and x-values as column indicies.}
\usage{
extreme_points_sparse(x)
}
\arguments{
\item{x}{a matrix, or vectr conatining measurment data row wise.}
}
\value{
a sparse matrix containing extrema information.
}
\description{
Finds kind and location of extrema and converts the to a sparse matrix, with measurment number as row indicies and x-values as column indicies.
}
\examples{
extreme_points_sparse(x)
}
