% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_SEC.R
\name{Plot_SEC}
\alias{Plot_SEC}
\title{Plot SEC}
\usage{
Plot_SEC(SEC, scan_dir)
}
\arguments{
\item{SEC}{a list of SEC matricies, obtained from Build_SEC()}

\item{scan_dir}{wich should be plotted}
}
\value{
a plot of SEC data
}
\description{
Plot SEC
}
\examples{
Plot_SEC(SEC, scan_dir = "reductive")
}
